Increase the size of program offsets to uint32_t
If a program exceeded the size of a unsigned short the GD would
either crash or get stuck in an infinite loop. Since it's
fairly easy to create such a program the offset size has been
increased to an uint32_t from an unsigned short.

More work is still required to support more strings.

Fixes #1