Implement DEFAULT_LC_CTYPE and DEFAULT_LC_ALL
Various parts of the driver are using locale dependant libc
methods which means that the locale must be set correctly for
things to work properly.

The default C locale does not allow the high bit ISO-8859-1 chars.

This is potentially breaking, to get the previous behaviour one
would set DEFAULT_LC_CTYPE to C.